fix(sonarcloud): resolve SonarCloud quality gate issues on new code
- Extract STATUS_ABIERTA, STATUS_CERRADA, STATUS_FINALIZADA constants in CallInteractor to eliminate duplicated string literals (S1192)
- Replace Stream.collect(Collectors.toList()) with Stream.toList() in CallInteractor, SaveProjectAdapter (S6204)
- Change responsibleId @RequestParam type from Integer to Long to avoid unnecessary unboxing in ProjectController (S2153)
- Add ZoneId.systemDefault() to all LocalDateTime.now() calls in ProjectProcedureAdapter, ResearchCallEntity, DocumentEntity, Membresia (S2171)
- Fill empty @PrePersist status-null block in ResearchCallEntity with default value (S108)
- Remove unused import CreateProjectRequest in ProjectMapper (S1128)
- Remove unused import java.util.Optional in ResearchCallModuleTest (S1128)
- Replace LocalDate.now() with fixed dates in ResearchCallModuleTest and ProjectModuleTest to avoid system clock in tests (S5977)
- Refactor assertThrows lambdas to have single throwing invocation in ResearchCallModuleTest (S5778)
- Replace no-arg lambdas with method references in ProjectModuleTest validateInvariants calls (S1612)

+++ fiis/src/main/java/com/sgi/fiis/proyectos/presentation/controller/ProjectController.java
@@ -45,12 +45,12 @@ public class ProjectController {
     @GetMapping
     @Operation(summary = "List research projects with optional filters")
     public ResponseEntity<List<ProjectResponse>> getProjects(
-            @RequestParam(value = "responsibleId", required = false) Integer responsibleId,
+            @RequestParam(value = "responsibleId", required = false) Long responsibleId,
             @RequestParam(value = "groupId", required = false) Integer groupId) {
         
         List<ProjectResponse> response;
         if (responsibleId != null) {
-            response = createProjectUseCase.getProjectsByResponsible(responsibleId.longValue());
+            response = createProjectUseCase.getProjectsByResponsible(responsibleId);
         } else if (groupId != null) {
             response = createProjectUseCase.getProjectsByGroup(groupId);
         } else {
